+++ R/g.analyse.R
@@ -289,6 +289,51 @@ g.analyse =  function(I, C, M, IMP, params_247 = c(), params_phyact = c(),
                             ID = ID,
                             fname = fname,
                             startt = startt)
+  
+  if (!is.null(M$QClog)) {
+    # Summarise the QC log (currently only expected from cwa Axivity files)
+    QCsummarise = function(QClog, wx) {
+      x = ifelse(test = length(wx) > 0,
+                 yes = sum(QClog$end[wx] - QClog$start[wx]) / 60,
+                 no = 0)
+      return(x)
+    }
+    # total imputation
+    impdone = which(M$QClog$imputed == TRUE)
+    file_summary$Dur_imputed = QCsummarise(M$QClog, impdone)
+    file_summary$Nblocks_imputed = length(impdone)
+    
+    # checksum
+    chsum_failed = which(M$QClog$checksum_pass == FALSE)
+    file_summary$Dur_chsum_failed = QCsummarise(M$QClog, chsum_failed)
+    file_summary$Nblocks_chsum_failed = length(chsum_failed)
+    
+    # nonincremental block ID
+    nonincremental = which(M$QClog$blockID_current - M$QClog$blockID_next != 1)
+    file_summary$Dur_nonincremental = QCsummarise(M$QClog, nonincremental)
+    file_summary$Nblocks_nonincremental = length(nonincremental)
+    
+    # sampling frequency issues
+    freqBlockHead = M$QClog$frequency_blockheader
+    frequency_bias = abs(M$QClog$frequency_observed - freqBlockHead) / freqBlockHead
+
+    freqissue = which(frequency_bias >= 0.05 & frequency_bias < 0.1)
+    file_summary$Dur_freqissue_5_10 = QCsummarise(M$QClog, freqissue)
+    file_summary$Nblock_freqissue_5_10 = length(freqissue)
+    
+    freqissue = which(frequency_bias >= 0.1 & frequency_bias < 0.2)
+    file_summary$Dur_freqissue_10_20 = QCsummarise(M$QClog, freqissue)
+    file_summary$Nblock_freqissue_10_20 = length(freqissue)
+    
+    freqissue = which(frequency_bias >= 0.2 & frequency_bias < 0.3)
+    file_summary$Dur_freqissue_20_30 = QCsummarise(M$QClog, freqissue)
+    file_summary$Nblock_freqissue_20_30 = length(freqissue)
+    
+    freqissue = which(frequency_bias >= 0.3)
+    file_summary$Dur_freqissue_30 = QCsummarise(M$QClog, freqissue)
+    file_summary$Nblock_freqissue_30 = length(freqissue)
+  }
+
   metrics_nav = list(lookat = lookat,
                     colnames_to_lookat = colnames_to_lookat,
                     longitudinal_axis_id = longitudinal_axis_id)