+++ NEWS.md
@@ -1,5 +1,8 @@
 # CHANGES IN GGIR VERSION 2.9-7
 
+- Part 1 + 2: File health log captured by dependency GGIRread::readAxivity 
+incorporated in data_quality_report #866
+
 - Part 1: Fix ID extraction for externally generated epoch files #869
 
 - Sleep log: Argument nnigths is now deprecated and number of nights are detected automatically in sleep logs #856